requirements.txt
asn1crypto==0.24.0
astroid==1.6.1
attrs==17.4.0
Automat==0.6.0
cffi==1.11.4
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
decorator==4.2.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.11.1
lazy-object-proxy==1.3.1
lxml==4.1.1
mccabe==0.6.1
parsel==1.3.1
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pylint==1.8.2
pyOpenSSL==17.5.0
pypandoc==1.4
python-slugify==1.2.4
queuelib==1.4.2
Scrapy==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==17.9.0
Unidecode==1.0.22
w3lib==1.19.0
wcwidth==0.1.7
wrapt==1.10.11
zope.interface==4.4.3
